% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fcum}
\alias{print.fcum}
\title{Print Method for fcum Objects}
\usage{
\method{print}{fcum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fcum}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for fcum Objects
}
