% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.fcum}
\alias{summary.fcum}
\title{Summary Method for fcum Objects}
\usage{
\method{summary}{fcum}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fcum}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Summary Method for fcum Objects
}
